+++ lib/loxxy/ast/all_lox_nodes.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative 'lox_fun_stmt'
+require_relative 'lox_this_expr'
 require_relative 'lox_variable_expr'
 require_relative 'lox_literal_expr'
 require_relative 'lox_noop_expr'